+++ src/main/java/de/markusbordihn/dogscompanion/commands/DogFollowCommand.java
@@ -32,6 +32,7 @@ import com.hypixel.hytale.server.npc.role.support.StateSupport;
 import de.markusbordihn.dogscompanion.Constants;
 import de.markusbordihn.dogscompanion.data.DogState;
 import de.markusbordihn.dogscompanion.manager.DogsManager;
+import java.util.Optional;
 import javax.annotation.Nonnull;
 
 final class DogFollowCommand extends DogCommand {
@@ -46,7 +47,7 @@ final class DogFollowCommand extends DogCommand {
   @Override
   protected void execute(
       @Nonnull CommandContext context, @Nonnull World world, @Nonnull Store<EntityStore> store) {
-    var entityRefOpt = getEntityFromArgument(this.entityArg, store, context);
+    Optional<Ref<EntityStore>> entityRefOpt = getEntityFromArgument(this.entityArg, store, context);
 
     if (entityRefOpt.isPresent()) {
       Ref<EntityStore> entityRef = entityRefOpt.get();