+++ src/main/java/de/markusbordihn/dogscompanion/commands/DogSleepCommand.java
@@ -31,6 +31,7 @@ import com.hypixel.hytale.server.npc.entities.NPCEntity;
 import de.markusbordihn.dogscompanion.Constants;
 import de.markusbordihn.dogscompanion.data.DogState;
 import de.markusbordihn.dogscompanion.manager.DogsManager;
+import java.util.Optional;
 import javax.annotation.Nonnull;
 
 final class DogSleepCommand extends DogCommand {
@@ -44,7 +45,7 @@ final class DogSleepCommand extends DogCommand {
   @Override
   protected void execute(
       @Nonnull CommandContext context, @Nonnull World world, @Nonnull Store<EntityStore> store) {
-    var entityRefOpt = getEntityFromArgument(this.entityArg, store, context);
+    Optional<Ref<EntityStore>> entityRefOpt = getEntityFromArgument(this.entityArg, store, context);
 
     if (entityRefOpt.isPresent()) {
       Ref<EntityStore> entityRef = entityRefOpt.get();