+++ src/main/java/de/markusbordihn/dogscompanion/data/DogState.java
@@ -28,6 +28,7 @@ public enum DogState {
   WANDERING,
   PLAYING,
   WAITING,
+  SEARCHING,
   ATTACKING,
   STRIKING,
   DEFENSE,